% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{checkReqPkgs}
\alias{checkReqPkgs}
\title{Check for installed package(s)}
\usage{
checkReqPkgs(pkgs, msg = "")
}
\arguments{
\item{pkgs}{Character vector with package(s) to check for.}

\item{msg}{Optionally additional text appended (with \code{paste0}) to the default error message.}
}
\value{
Returns error and message if not installed, otherwise \code{invisible(TRUE)}
}
\description{
Returns an error if required package(s) are not installed. Mostly used for checking whether packages listed under the Suggests field in the DESCRIPTION file are installed.
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
