% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmextract.R
\name{mmextract}
\alias{mmextract}
\title{Title}
\usage{
mmextract(mm, selection, min_length = 0, inverse = FALSE)
}
\arguments{
\item{mm}{}

\item{selection}{}

\item{min_length}{}

\item{inverse}{}
}
\description{
Title
}
