% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmextract_network.R
\name{mmextract_network}
\alias{mmextract_network}
\title{Title}
\usage{
mmextract_network(scaffolds, network, original_data, min_connections = 2,
  include_connections = "direct")
}
\arguments{
\item{scaffolds}{}

\item{network}{}

\item{original_data}{}

\item{min_connections}{}

\item{include_connections}{}
}
\description{
Title
}
