% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmheatmap.R
\name{mmheatmap}
\alias{mmheatmap}
\title{Heatmap of the most abundant bins}
\usage{
mmheatmap(mm, CLASSIFICATION = NULL, BIN_COL, TOPN = 20, tax_add = NULL)
}
\arguments{
\item{mm}{(\emph{required}) A dataframe loaded with \code{\link{mmload}}.}

\item{CLASSIFICATION}{Name of the variable in \code{mm} with taxonomic classification. The format must resemble that of the following: \code{d__Bacteria;p__Firmicutes_A;c__Clostridia;o__Tissierellales;f__Sedimentibacteraceae;g__Sedimentibacter;s__}. (\emph{Default: } \code{NULL})}

\item{BIN_COL}{(\emph{required}) Group the scaffolds by a variable in \code{mm}.}

\item{TOPN}{Number of the most abundant bins to display. (\emph{Default: } \code{20})}

\item{tax_add}{Taxonomic levels to show from the taxonomy variable, one of \code{c("Kingdom","Phylum","Class","Order","Family","Genus","Species")}. (\emph{Default: } \code{NULL})}
}
\value{
A ggplot object. Note that mmgenome2 hides all warnings produced by ggplot objects.
}
\description{
Generates a heatmap of the coverage profiles for the bins.
}
\examples{
library(mmgenome2)
data("mmgenome2")
mmgenome2
mmheatmap(
  mmgenome2,
  BIN_COL = "taxonomy",
  TOPN = 20
)
}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}

Rasmus Kirkegaard \email{rhk@bio.aau.dk}
}
