% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcircos.R
\name{mmlinks_filter}
\alias{mmlinks_filter}
\title{Title}
\usage{
mmlinks_filter(links, cov_sample, out_type, window_size = 2000,
  min_link = 3, min_cov = 5, cov_fprate = 1/50, end_length = NULL,
  remove_regions = NULL)
}
\arguments{
\item{links}{}

\item{cov_sample}{}

\item{out_type}{}

\item{window_size}{}

\item{min_link}{}

\item{min_cov}{}

\item{cov_fprate}{}

\item{end_length}{}

\item{remove_regions}{}
}
\description{
Title
}
