% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmload.R
\name{mmload}
\alias{mmload}
\title{mmload}
\usage{
mmload(assembly, coverage, essential_genes = NULL, kmer_pca = FALSE,
  kmer_BH_tSNE = FALSE, taxonomy = NULL, additional = NULL,
  verbose = TRUE, ...)
}
\arguments{
\item{assembly}{}

\item{coverage}{}

\item{essential_genes}{}

\item{kmer_pca}{}

\item{kmer_BH_tSNE}{}

\item{taxonomy}{}

\item{additional}{}

\item{verbose}{}

\item{...}{}
}
\description{
mmload
}
