% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmmerge.R
\name{mmmerge}
\alias{mmmerge}
\title{Merge mm object with more data}
\usage{
mmmerge(x, y, type)
}
\arguments{
\item{x}{mm object}

\item{y}{Object to merge with \code{x}}

\item{type}{Character string defining the type of data being merged}
}
\value{
A tibble
}
\description{
Internal function used in mmload. \code{y} can be a named vector, dataframe, or a list containing any of these types of data to be merged with \code{x}.
}
