% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmmerge_linkfiles.R
\name{mmmerge_linkfiles}
\alias{mmmerge_linkfiles}
\title{Load and merge files with mate-pair or paired-end connection information}
\usage{
mmmerge_linkfiles(path, verbose = TRUE)
}
\arguments{
\item{path}{(\emph{required}) Path to a folder.}

\item{verbose}{(\emph{Logical}) Whether to print status messages of the process. (\emph{Default: } \code{TRUE})}
}
\value{
A tibble
}
\description{
Searches a folder for all files with a filename ending with \code{"_link"}, and then all the files found are loaded and merged into one data frame. This is done by searching the files for all unique scaffold pairs and summing the number of connections per pair. The files must contain 3 columns called \code{"scaffold1"}, \code{"scaffold2"}, and \code{"connections"}.
}
\examples{
\dontrun{
  links <- mmmerge_linkfiles("path/to/folder")
}
}
