% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmnetwork.R
\name{mmnetwork}
\alias{mmnetwork}
\title{Network plot of mate-pair or paired-end connected scaffold reads}
\usage{
mmnetwork(mm, network, min_connections = 1, color_by = NULL,
  color_scale_log10 = FALSE, locator = FALSE, selection = NULL,
  highlight_labels = NULL, highlight_color = "darkred", links_scale = 1,
  scaffold_labels = FALSE, print_nolinks = FALSE, seed = 0)
}
\arguments{
\item{mm}{(\emph{required}) A dataframe loaded with \code{\link{mmload}}.}

\item{network}{(\emph{required}) Paired-end or mate-pair connections between scaffolds in long format. The first and second columns must contain all connected scaffold pairs and the third column the number of connections.}

\item{min_connections}{Filter all scaffold pairs with equal to or less than this number of connections. (\emph{Default: } \code{2})}

\item{color_by}{Color the scaffolds by a variable in \code{mm}. (\emph{Default: } \code{NULL})}

\item{color_scale_log10}{(\emph{Logical}) Log10-scale the color gradient when \code{color_by} is set and the variable is continuous (\emph{Default: } \code{FALSE})}

\item{locator}{(\emph{Logical}) When \code{TRUE}, left-clicks in the plot are captured and the exact x/y-coordinates of the mouse clicks are captured and returned. These coordinates can be used to highlight a selection of scaffolds in the plot, and to extract all scaffolds within the selection. (\emph{Default: } \code{FALSE})}

\item{selection}{A 2-column dataframe with the x and y coordinates of points with which to draw a polygon onto the plot to highlight a selected region. A selection can be obtained by using the locator feature (by \code{locator = TRUE}). (\emph{Default: } \code{NULL})}

\item{highlight_labels}{A dataframe or vector of scaffold names whose labels to highlight in the plot (colored by \code{highlight_color}). (\emph{Default: } \code{NULL})}

\item{highlight_color}{The color with which to highlight the scaffold labels defined by \code{highlight_labels}. (\emph{Default: } \code{"darkred"})}

\item{links_scale}{A factor to scale the sizes of the links plotted between scaffolds. (\emph{Default: } \code{1})}

\item{scaffold_labels}{Add labels with the scaffold names of \emph{all} scaffolds.}

\item{print_nolinks}{(\emph{Logical}) Print the names of all scaffolds with no links to other scaffolds to the console. (\emph{Default: } \code{FALSE})}

\item{seed}{Network plots are based on Random Number Generation, and this is used to set a specific random seed (with \code{set.seed}) allowing for reproducible network plots. Set to NULL to generate a unique network plot with each run of \code{mmnetwork}. (\emph{Default: } \code{0})}
}
\value{
A ggplot2 object.
}
\description{
Plots all connected scaffolds in a network plot. Scaffolds can then be highlighted and extracted using the locator and selection features.
}
\section{Using the locator with mmnetwork}{

Unlike a plot generated by \code{\link{mmplot}}, the coordinates of points in each dimension in a \code{\link{mmnetwork}} plot is not contained within the provided dataframe \code{mm} itself. As such \code{\link{mmnetwork}} does not have a corresponding extract function like \code{\link{mmextract}} is for \code{\link{mmplot}}. Instead, a subset of \code{mm} containing all scaffolds within the selection polygon is available in the returned ggplot object with \code{plot$data_in_selection}.
}

\examples{
library(mmgenome2)
data(mmgenome2)
data(paired_ends)
mmgenome2
selection <- data.frame(cov_C13.11.25 = c(7.2, 16.2, 25.2, 23.3, 10.1),
                        cov_C14.01.09 = c(47, 77, 52.8, 29.5, 22.1))
mmgenome2_extraction <- mmextract(mmgenome2, selection = selection)
mmgenome2_extraction
p <- mmnetwork(mmgenome2_extraction, 
               network = paired_ends, 
               min_connections = 1,
               color_by = "taxonomy")
p

}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}

Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
