% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot.R
\name{mmplot}
\alias{mmplot}
\title{plot}
\usage{
mmplot(mm, x, y, min_length = 0, locator = FALSE, selection = NULL,
  x_scale = NULL, x_limits = NULL, y_scale = NULL, y_limits = NULL,
  color_by = NULL, alpha = 0.1, scaffold_labels = FALSE,
  fixed_size = NULL, size_scale = 1, duplicate_genes = FALSE)
}
\arguments{
\item{mm}{(\emph{required}) A dataframe loaded with \code{\link{mmload}}.}

\item{x}{(\emph{required}) The variable from \code{mm} to plot on the first axis.}

\item{y}{(\emph{required}) The variable from \code{mm} to plot on the second axis.}

\item{min_length}{Remove scaffolds with a length at or below this size before plotting. (\emph{Default: } \code{0})}

\item{locator}{(\emph{Logical}) . (\emph{Default: } \code{FALSE})}

\item{selection}{A 2-column dataframe with the x and y coordinates of points with which to draw a polygon onto the plot to highlight selected regions.}

\item{x_scale}{Apply a log10-scale (\code{"log10"}) or a square-root scale \code{"sqrt"} to the x axis. (\emph{Default: } \code{NULL})}

\item{x_limits}{Axis limits of the x axis. (\emph{Default: } \code{NULL})}

\item{y_scale}{Apply a log10-scale (\code{"log10"}) or a square-root scale \code{"sqrt"} to the y axis. (\emph{Default: } \code{NULL})}

\item{y_limits}{Axis limits of the y axis. (\emph{Default: } \code{NULL})}

\item{color_by}{(\emph{Default: } \code{NULL})}

\item{alpha}{(\emph{Default: } \code{0.1})}

\item{scaffold_labels}{(\emph{Logical})  (\emph{Default: } \code{FALSE})}

\item{fixed_size}{(\emph{Default: } \code{NULL})}

\item{size_scale}{(\emph{Default: } \code{1})}

\item{duplicate_genes}{(\emph{Logical})  (\emph{Default: } \code{FALSE})}
}
