% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcircos.R
\name{mmplot_circos}
\alias{mmplot_circos}
\title{Title}
\usage{
mmplot_circos(assembly = get("assembly", envir = globalenv()), coverage,
  profile = NULL, links = NULL, assembly_col = NULL, gc = TRUE,
  line_width = 0.5, window_size = 5000, print_pdf = NULL)
}
\arguments{
\item{assembly}{}

\item{coverage}{}

\item{profile}{}

\item{links}{}

\item{assembly_col}{}

\item{gc}{}

\item{line_width}{}

\item{window_size}{}

\item{print_pdf}{}
}
\description{
Title
}
