% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot_cov_profiles.R
\name{mmplot_cov_profiles}
\alias{mmplot_cov_profiles}
\title{Plot all coverage profiles}
\usage{
mmplot_cov_profiles(mm, normalise = TRUE, alpha = 0.6, color_by = NULL,
  color_vector = c("blue", "green", "red"), color_scale_log10 = FALSE,
  y_scale_log10 = FALSE, plot_lines = TRUE)
}
\arguments{
\item{mm}{}

\item{normalise}{}

\item{alpha}{}

\item{color_by}{}

\item{color_vector}{}

\item{color_scale_log10}{}

\item{y_scale_log10}{}

\item{plot_lines}{}
}
\value{
A ggplot2 object.
}
\description{
To look for patterns... WIP
}
\examples{
library(mmgenome2)
data(mmgenome2)
mmgenome2
selection <- data.frame(cov_C13.12.03 = c(7.676, 5.165, 6.386, 10.933), 
                        cov_C14.01.09 = c(24.852, 32.545, 53.062, 38.52))
mmgenome2_extraction <- mmextract(mmgenome2, 
                                  min_length = 3000,
                                  selection = selection,
                                  inverse = FALSE)
mmplot_cov_profiles(mmgenome2_extraction,
                    color_by = "taxonomy")
}
