% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot_pairs.R
\name{mmplot_pairs}
\alias{mmplot_pairs}
\title{Title}
\usage{
mmplot_pairs(mm, variables, textsize = 8, log = NULL)
}
\arguments{
\item{mm}{}

\item{variables}{}

\item{textsize}{}

\item{log}{}
}
\description{
Title
}
