% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmscanbins.R
\name{mmscanbins}
\alias{mmscanbins}
\title{Retrieve scaffold names from all genome bins in a folder}
\usage{
mmscanbins(binfolder, namesOnly = TRUE)
}
\arguments{
\item{binfolder}{(\emph{required}) Path to the folder to scan.}

\item{namesOnly}{Return only the scaffold names (\code{TRUE}) or a complete list with the DNA sequences of the scaffolds in each file (\code{FALSE}). (\emph{Default: } \code{TRUE})}
}
\value{
A character vector (if \code{namesOnly = TRUE}) or a list (if \code{namesOnly = FALSE}).
}
\description{
Reads all FASTA files (genome bins) in a folder and returns all unique scaffold names. This is useful to get an overview of which scaffolds are already extracted from the metagenome by for example highlighting them in plots.
}
\examples{
\dontrun{
#Use mmextract and mmexport to extract scaffolds from the assembly and 
#write their sequences to a FASTA file per bin. Then the mmscanbins() function 
#can be used to retrieve the names of all scaffolds already extracted, and 
#then for example highlight them in a plot:

binned_scaffolds <- mmscanbins("path/to/folder", namesOnly = TRUE)
binned_scaffolds
mmplot(mmgenome2,
       min_length = 10000,
       x = "cov_C13.12.03",
       y = "cov_C14.01.09",
       color_by = "taxonomy",
       highlight_scaffolds = binned_scaffolds,
       highlight_color = "darkred",
       x_scale = "log10",
       y_scale = "log10")
}
}
