% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcircos.R
\name{mmsynteny_align}
\alias{mmsynteny_align}
\title{Title}
\usage{
mmsynteny_align(bin1, bin2, synteny_filtered, window_size = 500,
  out_type = "id_windows", min_avg_id = 0.7, min_length = 1000)
}
\arguments{
\item{bin1}{}

\item{bin2}{}

\item{synteny_filtered}{}

\item{window_size}{}

\item{out_type}{}

\item{min_avg_id}{}

\item{min_length}{}
}
\description{
Title
}
