% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmcircos.R
\name{mmsynteny_filter}
\alias{mmsynteny_filter}
\title{Title}
\usage{
mmsynteny_filter(synteny, min_score, min_length = 1000)
}
\arguments{
\item{synteny}{}

\item{min_score}{}

\item{min_length}{}
}
\value{

}
\description{

}
