% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paired_ends}
\alias{paired_ends}
\title{Example data}
\format{
A data frame
}
\description{
Very little info here
}
\examples{
\dontrun{
data("paired_ends")
View(paired_ends)
}
}
\keyword{data}
