% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ggplot.R
\name{print.ggplot}
\alias{print.ggplot}
\title{Custom ggplot object print to suppress warnings}
\usage{
\method{print}{ggplot}(...)
}
\arguments{
\item{...}{All arguments are directly passed to \code{ggplot2:::print.ggplot()}}
}
\description{
Simply wraps \code{ggplot2:::print.ggplot()} in \code{suppressWarnings()}. The main reason for this is to hide warnings produced by log10-transformation in \code{mmplot}.
}
