% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmexpand_network.R
\name{mmexpand_network}
\alias{mmexpand_network}
\title{Find}
\usage{
mmexpand_network(scaffolds, network, original_data, min_connections = 2,
  include_connections = "direct")
}
\arguments{
\item{scaffolds}{}

\item{network}{}

\item{original_data}{}

\item{min_connections}{}

\item{include_connections}{}
}
\description{
Find
}
