% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmexport.R
\name{mmexport}
\alias{mmexport}
\title{Export DNA sequences of selected scaffolds}
\usage{
mmexport(
  scaffolds,
  assembly = get("assembly", envir = .GlobalEnv),
  file = "exported_assembly.fa"
)
}
\arguments{
\item{scaffolds}{(\emph{required}) Either a vector of scaffold names or a dataframe loaded with \code{\link{mmload}} containing the scaffolds to extract from the \code{assembly}.}

\item{assembly}{(\emph{required}) The assembly to subset and export. Must be an object of class \code{"DNAStringSet"} as loaded with \code{\link{readDNAStringSet}}. By default any object named "assembly" in the global environment will be used.}

\item{file}{The file path and file name of the file to write the sequences to. (\emph{Default: } \code{"exported_assembly.fa"})}
}
\description{
Writes the DNA sequences of a given set of scaffolds from the assembly to a FASTA file.
}
\examples{
library(mmgenome2)
data(mmgenome2)
mmgenome2
selection <- data.frame(
  cov_C13.11.25 = c(7.2, 16.2, 25.2, 23.3, 10.1),
  cov_C14.01.09 = c(47, 77, 52.8, 29.5, 22.1)
)
mmgenome2_extraction <- mmextract(mmgenome2,
  min_length = 3000,
  selection = selection,
  inverse = FALSE
)
mmgenome2_extraction
\dontrun{
mmexport(mmgenome2_extraction,
  assembly = assembly,
  file = "bins/exported_assembly.fa"
)
}

}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}

Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
