% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmload.R
\name{mmload}
\alias{mmload}
\title{Load and combine data for use with other mmgenome2 functions}
\usage{
mmload(assembly, coverage, essential_genes = NULL, taxonomy = NULL,
  additional = NULL, kmer_pca = FALSE, kmer_BH_tSNE = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{assembly}{(\emph{required}) A character string with the path to the assembly FASTA file, or the assembly as already loaded with \code{\link[Biostrings]{readDNAStringSet}}.}

\item{coverage}{(\emph{required}) A \code{vector}, \code{dataframe}, or a \code{list} hereof containing coverage of each scaffold. The prefix \code{"cov_"} will be appended to all coverage column names in the output.
\describe{
  \item{\code{vector}}{If provided as a vector, the elements of the vector must be named by the scaffold names exactly matching those of the assembly.}
  \item{\code{dataframe}}{If provided as a dataframe, the first column must contain the scaffold names exactly matching those of the assembly, and any additional column(s) contain coverage of each scaffold.}
  \item{\code{list}}{If provided as a list, it must contain any number of \code{vector} or \code{dataframe} objects as described above.}
}}

\item{essential_genes}{A 2-column dataframe with scaffold names in the first column and gene ID's in the second. Can contain duplicates. (\emph{Default: } \code{NULL})}

\item{taxonomy}{A dataframe containing taxonomy assigned to the scaffolds. The first column must contain the scaffold names. (\emph{Default: } \code{NULL})}

\item{additional}{A dataframe containing any additional data. The first column must contain the scaffold names. (\emph{Default: } \code{NULL})}

\item{kmer_pca}{(\emph{Logical}) Perform Principal Components Analysis of tetranucleotide frequencies of each scaffold and merge the scores of the 3 most significant axes. (\emph{Default: } \code{FALSE})}

\item{kmer_BH_tSNE}{(\emph{Logical}) Calculate Barnes-Hut t-Distributed Stochastic Neighbor Embedding (B-H t-SNE) representations of tetranucleotide frequencies using \code{\link[Rtsne.multicore]{Rtsne.multicore}} and merge the result. Additional arguments may be required for success, refer to the documentation of \code{\link[Rtsne.multicore]{Rtsne.multicore}}. (\emph{Default: } \code{FALSE})}

\item{verbose}{(\emph{Logical}) Whether to print status messages during the loading process. (\emph{Default: } \code{TRUE})}

\item{...}{Additional arguments are passed on to \code{\link[Rtsne.multicore]{Rtsne.multicore}}.}
}
\value{
A dataframe (tibble) compatible with other mmgenome2 functions.
}
