% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmload.R
\name{mmload}
\alias{mmload}
\title{Load and combine data for use with other mmgenome2 functions}
\usage{
mmload(assembly, coverage, essential_genes = NULL, taxonomy = NULL,
  additional = NULL, kmer_pca = FALSE, kmer_BH_tSNE = FALSE,
  verbose = TRUE, ...)
}
\arguments{
\item{assembly}{(\emph{required}) A character string with the path to the assembly FASTA file, or the assembly as already loaded with \code{\link{readDNAStringSet}}.}

\item{coverage}{(\emph{required}) A path to a folder to scan for coverage files, or otherwise a \code{vector}, \code{data.frame}, or a \code{list} hereof containing coverage of each scaffold. The prefix \code{"cov_"} will be appended to all coverage column names in the output so that \code{\link{mmstats}} knows which columns are coverage columns.
\describe{
  \item{\code{vector}}{If provided as a vector, the elements of the vector must be named by the scaffold names exactly matching those of the assembly.}
  \item{\code{data.frame}}{If provided as a dataframe, the first column must contain the scaffold names exactly matching those of the assembly, and any additional column(s) contain coverage of each scaffold.}
  \item{\code{list}}{If provided as a list, it must contain any number of \code{vector}'s or \code{data.frame}'s as described above. If names are assigned to the objects in the list, then they will be used as column names in the output (does not apply to any dataframes that may have more than 2 columns, however).}
  \item{\code{path}}{If a path to a folder is provided, then all files with filenames ending with \code{"_cov"} will be loaded (by the \code{\link[data.table]{fread}} function) into a list of \code{data.frame}'s and treated as if a \code{list} of \code{data.frame}'s were provided. The filenames (stripped from extension and \code{"_cov"}) will then be used as column names in the output. \strong{Note: only the first 2 columns will be used in the loaded files!}}
}}

\item{essential_genes}{Either a path to a delimited file (with a 1-space delimiter: " ") containing the essential genes, or a 2-column dataframe with scaffold names in the first column and gene ID's in the second. Can contain duplicates. (\emph{Default: } \code{NULL})}

\item{taxonomy}{A dataframe containing taxonomy assigned to the scaffolds. The first column must contain the scaffold names. (\emph{Default: } \code{NULL})}

\item{additional}{A dataframe containing any additional data. The first column must contain the scaffold names. (\emph{Default: } \code{NULL})}

\item{kmer_pca}{(\emph{Logical}) Perform Principal Components Analysis of tetranucleotide frequencies of each scaffold and merge the scores of the 3 most significant axes. (\emph{Default: } \code{FALSE})}

\item{kmer_BH_tSNE}{(\emph{Logical}) Calculate Barnes-Hut t-Distributed Stochastic Neighbor Embedding (B-H t-SNE) representations of tetranucleotide frequencies using \code{\link[Rtsne.multicore]{Rtsne.multicore}} and merge the result. Additional arguments may be required for success, refer to the documentation of \code{\link[Rtsne.multicore]{Rtsne.multicore}}. (\emph{Default: } \code{FALSE})}

\item{verbose}{(\emph{Logical}) Whether to print status messages during the loading process. (\emph{Default: } \code{TRUE})}

\item{...}{Additional arguments are passed on to \code{\link[Rtsne.multicore]{Rtsne.multicore}}.}
}
\value{
A dataframe (tibble) compatible with other mmgenome2 functions.
}
\description{
Loads, validates and combines multiple aspects of metagenome data into one dataframe for use with all mmgenome2 functions, including scaffold assembly sequences, scaffold coverage, essential genes, taxonomy, and more.
}
\examples{
\dontrun{
  library(mmgenome2)   
  mm <- mmload(
    assembly = "path/to/assembly.fa",
    coverage = list(nameofcoverage1 = read.csv("path/to/coveragetable1.csv", col.names = TRUE),
                    nameofcoverage2 = read.csv("path/to/coveragetable2.csv", col.names = TRUE)),
    essential_genes = "path/to/ess_genes.txt",
    verbose = TRUE
  )
  mm
}

}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}
}
