% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmnetwork.R
\name{mmnetwork}
\alias{mmnetwork}
\title{Title}
\usage{
mmnetwork(mm, network, min_connections = 2, color_by = NULL,
  locator = FALSE, selection = NULL, highlight_labels = NULL,
  highlight_color = "darkred", color_scale_log10 = FALSE, links_scale = 1,
  scaffold_labels = FALSE, print_nolinks = FALSE, seed = .mm_Random.seed)
}
\arguments{
\item{mm}{(\emph{required}) A dataframe loaded with \code{\link{mmload}}.}

\item{network}{}

\item{min_connections}{}

\item{color_by}{}

\item{locator}{}

\item{selection}{}

\item{highlight_labels}{}

\item{highlight_color}{}

\item{color_scale_log10}{}

\item{links_scale}{}

\item{scaffold_labels}{}

\item{print_nolinks}{}

\item{seed}{}
}
\description{
Title
}
