% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot.R
\name{mmplot}
\alias{mmplot}
\title{mmplot}
\usage{
mmplot(mm, x, y, color_by = NULL, alpha = 0.1, fixed_size = NULL,
  size_scale = 1, min_length = 0)
}
\arguments{
\item{mm}{}

\item{x}{}

\item{y}{}

\item{color_by}{}

\item{alpha}{}

\item{fixed_size}{}

\item{size_scale}{}

\item{min_length}{}
}
\description{
mmplot
}
