% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot.R
\name{mmplot}
\alias{mmplot}
\title{Visualise metagenomes in various ways}
\usage{
mmplot(mm, x, y, min_length = 0, color_by = NULL, locator = FALSE,
  selection = NULL, x_scale = NULL, x_limits = NULL, y_scale = NULL,
  y_limits = NULL, alpha = 0.1, scaffold_labels = FALSE,
  fixed_size = NULL, size_scale = 1, shared_genes = FALSE)
}
\arguments{
\item{mm}{(\emph{required}) A dataframe loaded with \code{\link{mmload}}.}

\item{x}{(\emph{required}) The variable from \code{mm} to plot on the first axis.}

\item{y}{(\emph{required}) The variable from \code{mm} to plot on the second axis.}

\item{min_length}{Remove scaffolds with a length at or below this value before plotting. (\emph{Default: } \code{0})}

\item{color_by}{Color the scaffolds by a variable in \code{mm}. (\emph{Default: } \code{NULL})}

\item{locator}{(\emph{Logical}) When \code{TRUE}, left-clicks in the plot are captured and the exact x/y-coordinates of the mouse clicks are returned. These coordinates can be used to highlight a selection of scaffolds in the plot, and also be used with \code{\link{mmextract}} to extract all scaffolds within the selection from the data. (\emph{Default: } \code{FALSE})}

\item{selection}{A 2-column dataframe with the x and y coordinates of points with which to draw a polygon onto the plot to highlight a selected region. A selection can be obtained by using the locator feature (by \code{locator = TRUE}). (\emph{Default: } \code{NULL})}

\item{x_scale}{Log10-scale (\code{"log10"}) or square-root scale \code{"sqrt"} the x axis. (\emph{Default: } \code{NULL})}

\item{x_limits}{Axis limits of the x axis. (\emph{Default: } \code{NULL})}

\item{y_scale}{Log10-scale (\code{"log10"}) or square-root scale \code{"sqrt"} the y axis. (\emph{Default: } \code{NULL})}

\item{y_limits}{Axis limits of the y axis. (\emph{Default: } \code{NULL})}

\item{alpha}{The transparancy of the scaffold points, where 0 is invisible and 1 is opaque. (\emph{Default: } \code{0.1})}

\item{scaffold_labels}{Add labels of a selection of scaffolds by providing either a character vector of scaffold names, or a dataframe with scaffold names in the first column. If set to \code{TRUE} then \emph{all} scaffolds will be labelled. (\emph{Default: } \code{FALSE})}

\item{fixed_size}{A fixed size for all scaffolds if set. If \code{NULL} then the scaffolds are scaled by length. (\emph{Default: } \code{NULL})}

\item{size_scale}{A factor to scale the sizes of the scaffolds plotted. Only applies when \code{fixed_size} is set to \code{NULL} and the scaffolds are scaled by length. (\emph{Default: } \code{1})}

\item{shared_genes}{(\emph{Logical}) If \code{TRUE}, lines will be drawn between scaffolds with any shared gene(s). (\emph{Default: } \code{FALSE})}
}
\value{
A ggplot2 object.
}
\description{
Plots any information about the scaffolds contained in the given mm object as a scatterplot, for example different coverage variables, scaffold length, GC content, or anything else that were loaded with \code{\link{mmload}}. Scaffolds can then be highlighted and extracted using the locator and selection features.
}
\examples{
library(mmgenome2)
data(mmgenome2)
mmgenome2
mmplot(mmgenome2,
       min_length = 10000,
       x = "cov_C13.12.03",
       y = "cov_C14.01.09",
       color_by = "phylum",
       #locator = TRUE,
       x_scale = "log10",
       y_scale = "log10")
#Set "locator = TRUE" to interactively capture the coordinates of 
#mouse clicks in an mmplot, or provide coordinates with "selection":
selection <- data.frame(cov_C13.12.03 = c(7.676, 5.165, 6.386, 10.933), 
                        cov_C14.01.09 = c(24.852, 32.545, 53.062, 38.52))
mmplot(mmgenome2,
       min_length = 10000,
       x = "cov_C13.12.03",
       y = "cov_C14.01.09",
       color_by = "phylum",
       x_scale = "log10",
       y_scale = "log10",
       x_limits = c(1,NA), #zoom in at minimum 1x coverage
       y_limits = c(1,NA), #zoom in at minimum 1x coverage
       selection = selection) #highlight the selection marked with locator

}
\author{
Kasper Skytte Andersen \email{ksa@bio.aau.dk}

Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
