% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmplot_pairs.R
\name{mmplot_pairs}
\alias{mmplot_pairs}
\title{Generates a ggplot2 style pairs plot}
\usage{
mmplot_pairs(mm, variables)
}
\arguments{
\item{mm}{(required) The dataframe to plot.}

\item{variables}{(required) The vaiables to plot as a vector.}

\item{textsize}{Size of the legend text (deafult = 8).}

\item{...}{Arguments passed on to mmplot()}
}
\value{
A pairs plot object.
}
\description{
Generates a ggplot2 style pairs plot.
}
\examples{
data(rocco)

mmplot_pairs(mm = d, variables = c("C13.12.03", "C14.01.09", "C13.11.25", "gc", "PC2"), color = "essential", log = c("C13.12.03", "C14.01.09", "C13.11.25"), minlength = 10000)


}
\author{
Soren M. Karst \email{smk@bio.aau.dk}

Mads Albertsen \email{MadsAlbertsen85@gmail.com}
}
